+++ tests/test_metadata.py
@@ -46,7 +46,7 @@ class TestMetadata:
         )
         assert __version__ == PYCLN_METADATA["version"]
 
-    @pytest.mark.skipif(  # TODO: Remove `-` sign when alpha.0.
+    @pytest.mark.skipif(  # TODO: Remove `-` sign when 0.0.1.
         getenv("CI", "false") != "-true" or not VersionInfo.isvalid(__version__),
         reason="Invalid semantic-version.",
     )